update messages structure
