+++ demo/actions.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from django.template.loader import render_to_string
 from django.conf import settings
 from telegram_framework import actions
-from telegram_framework import messages, media
+from telegram_framework import messages
 from .models import create_info_text
 
 
@@ -37,5 +37,5 @@ def render_template_example(bot, message):
 def load_picture_example(bot, message):
     file_path = Path(settings.BASE_DIR) / 'static' / 'logo_1280_640.png'
     caption = messages.create_message('<b>DTF</b> LOGO', bot, format_type='HTML')
-    image = media.create_image(bot, file_path, caption)
+    image = messages.create_image(bot, file_path, caption)
     return actions.send_image(message.chat, image)
