+++ telegram_framework/messages/__init__.py
@@ -1,4 +1,5 @@
-from .message import Message, create_message
+from .text_message import Message, create_message
 from .reply import Reply, create_reply
 from .chat_message import ChatMessage, create_chat_message
 from .chat_message_base import is_chat_message
+from .media import Image, create_image
