+++ telegram_framework/messages/core/__init__.py
@@ -1,0 +1,2 @@
+from .core_message_base import CoreMessageBase
+from .core_message_default import CoreMessageDefault
