+++ telegram_framework/messages/core/core_message_base.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+from typing import Any
+
+
+@dataclass(frozen=True)
+class CoreMessageBase:
+    sender: Any
