+++ telegram_framework/messages/media/image_base.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+from telegram_framework.messages.core import CoreMessageBase
+
+
+@dataclass(frozen=True)
+class ImageBase(CoreMessageBase):
+    file_path: str
