+++ telegram_framework/messages/media/image_default.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+from telegram_framework.messages import Message
+from telegram_framework.messages.core import CoreMessageDefault
+
+
+@dataclass(frozen=True)
+class ImageDefault(CoreMessageDefault):
+    caption: Message = None
