--- telegram_framework/messages/message_base.py
@@ -1,13 +1,0 @@
-from dataclasses import dataclass
-from typing import Any
-
-
-@dataclass(frozen=True)
-class MessageToSend:
-    sender: Any
-
-
-@dataclass(frozen=True)
-class MessageBase(MessageToSend):
-    text: str
-    # sender: Any
