+++ telegram_framework/messages/reply/reply.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass
 from typing import Any
-from .errors import MessageNotInChatError
-from .message_default import MessageDefault
+from telegram_framework.messages.errors import MessageNotInChatError
+from telegram_framework.messages.text_message import MessageDefault
 from .reply_base import ReplyBase
-from .chat_message_base import is_chat_message
+from telegram_framework.messages.chat_message_base import is_chat_message
 
 
 @dataclass(frozen=True)
