+++ telegram_framework/messages/text_message/message_base.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+from telegram_framework.messages.core import CoreMessageBase
+
+@dataclass(frozen=True)
+class MessageBase(CoreMessageBase):
+    text: str
